Feat: Add level loading menu
LevelSelectionUI can have several LevelSelectorUI.

If a LevelSelectorUI is clicked it calls load_level() on the ScenesManger. This function loads Game and adds a Level as a child of Game. Game also has its var 'level' set to this Level node. Stats wait for all of this to be setup and then updates its node refs.